cl-gfl3.lisp: (create-window): discern window_hints_hard Hard and Soft constraints
modifiy CREATE-WINDOW behaviour to only call glfwWindowHint on the
hints: :STEREO, :OPENGL-FORWARD-COMPAT, :CLIENT-API, and
:OPENGL-PROFILE, (the "hard constraint" hints) if and only if they
are explicitly supplied in the call to create-window.

info from the GLFW3 "Window Guide" doc
```
Some window hints are hard constraints.  These must match the available
capabilities _exactly_ for window and context creation to succeed.  Hints
that are not hard constraints are matched as closely as possible, but the
resulting context and framebuffer may differ from what these hints
requested.
```

specifying an incorrect (i.e. not supported) value for hard constraint
hint on a particular system where glfw3 runs can result in an unusable
context being created.  So we take care not to specify any defaults
for these hints, but defer to the caller to specify any values.

we do this by modifying the keyword lambda list to include a
<hint>-supplied-p binding, and check that before setting the hint.

todo backward compatibility: CLIENT-API would default to OPENGL-API if
not supplied.